GeoLocation/requirements.txt
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.17
django-import-export==2.5.0
et-xmlfile==1.0.1
geographiclib==1.50
geopy==2.1.0
gis==0.2.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
